Basic ("the console boots and functions as expected") 22.0.0 support

+++ libraries/libmesosphere/source/kern_initial_process.cpp
@@ -285,7 +285,7 @@ namespace ams::kern {
             MESOSPHERE_INIT_ABORT_UNLESS(expected_size != 0);
 
             /* Ensure that the size we need to reserve is as we expect it to be. */
-            const size_t total_size = util::AlignUp(g_initial_process_binary_header.size, PageSize);
+            const u32 total_size = util::AlignUp(g_initial_process_binary_header.size, PageSize);
             MESOSPHERE_ABORT_UNLESS(total_size == expected_size);
             MESOSPHERE_ABORT_UNLESS(total_size <= InitialProcessBinarySizeMax);
